Rename point_set to set
