+++ posemir_discovery/src/sia.rs
@@ -5,7 +5,7 @@
 use crate::algorithm::MtpAlgorithm;
 use crate::point_set::mtp::Mtp;
 use crate::point_set::point::Point;
-use crate::point_set::point_set::PointSet;
+use crate::point_set::set::PointSet;
 use crate::utilities::sort;
 
 /// Implements the SIA algorithm [Meredith et al. 2002].
@@ -87,7 +87,7 @@ mod tests {
     use crate::point_set::mtp::Mtp;
     use crate::point_set::pattern::Pattern;
     use crate::point_set::point::Point2Df64;
-    use crate::point_set::point_set::PointSet;
+    use crate::point_set::set::PointSet;
     use crate::sia::Sia;
 
     const ALGORITHM: Sia = Sia {};
