+++ posemir_discovery/src/siatec_ch.rs
@@ -13,7 +13,7 @@ use crate::algorithm::TecAlgorithm;
 use crate::point_set::mtp::Mtp;
 use crate::point_set::pattern::Pattern;
 use crate::point_set::point::Point;
-use crate::point_set::point_set::PointSet;
+use crate::point_set::set::PointSet;
 use crate::point_set::tec::Tec;
 use crate::siatec_c::SiatecC;
 
@@ -292,7 +292,7 @@ mod tests {
     use crate::algorithm::TecAlgorithm;
     use crate::point_set::pattern::Pattern;
     use crate::point_set::point::Point2Df64;
-    use crate::point_set::point_set::PointSet;
+    use crate::point_set::set::PointSet;
     use crate::point_set::tec::Tec;
     use crate::siatec_c::SiatecC;
     use crate::siatec_ch::SiatecCH;
