+++ posemir_discovery/src/siatec_compress.rs
@@ -9,7 +9,7 @@ use crate::algorithm::TecAlgorithm;
 use crate::heuristic::{stats_of, TecStats};
 use crate::point_set::pattern::Pattern;
 use crate::point_set::point::Point;
-use crate::point_set::point_set::PointSet;
+use crate::point_set::set::PointSet;
 use crate::point_set::tec::Tec;
 
 /// Implements the SIATECCompress algorithm as described in [Meredith2013].
@@ -108,7 +108,7 @@ mod tests {
     use crate::algorithm::TecAlgorithm;
     use crate::point_set::pattern::Pattern;
     use crate::point_set::point::Point2Df64;
-    use crate::point_set::point_set::PointSet;
+    use crate::point_set::set::PointSet;
     use crate::siatec::Siatec;
     use crate::siatec_compress::SiatecCompress;
